[autotuner] bf16/fp16 reduction seed: CE wide-V w8 win + jsd correctness fix
Two changes to the reduction seed heuristic for half precision:

1. jsd bf16/fp16 correctness (examples/jsd.py): intermediate_dX was declared at
   _input.dtype but the loop's exp/log math promotes to fp32 -> ControlFlowTensorMismatch
   at bf16/fp16. Set it to fp32 (matches the sibling intermediate_loss + the fp32 dX
   output). True fp32 no-op (codegen byte-identical).

2. cross_entropy bf16 wide-V num_warps=8 (+~49% on V=32k/50k, GPT2/Llama2 vocabs) via a
   new faithful fact `full_width_output` (device_ir._has_full_width_output_store: does a
   store write the result back over the reduction extent [M,N] vs a per-row scalar [M]).
   Gate: rnumel>16384 AND row_reread AND NOT full_width_output AND size_hint*itemsize
   <=100KiB -> w8. ncu mechanism: a re-read scalar-output reduction is reduction-tree-bound,
   so w32's cross-warp shared-mem tree throttles reads; fewer warps win. CUDA-graph-validated
   M-stable, zero-regression: fp32 untouched (2x bytes > cap), and full-width (layer_norm/
   softmax/welford/rms) + non-reread (sum/long_sum/kl/jsd) kernels keep w32. Survived
   Gate A+D 3/3. fp32 seed configs provably unchanged (config-invariant vs baseline).

+++ examples/jsd.py
@@ -99,7 +99,11 @@ def jsd_forward(
                     dX[tile_bt, tile_X] = 0.0
                 continue
         intermediate_loss = hl.zeros([tile_bt, block_size_n], dtype=torch.float32)
-        intermediate_dX = hl.zeros([tile_bt, block_size_n], dtype=_input.dtype)
+        # Accumulate dX in fp32 (the loop body's exp/log math promotes to fp32, so a
+        # _input.dtype accumulator mismatches the carried value at bf16/fp16 ->
+        # ControlFlowTensorMismatch; fp32 also matches intermediate_loss, the fp32 dX
+        # output, and is a no-op at fp32 input).
+        intermediate_dX = hl.zeros([tile_bt, block_size_n], dtype=torch.float32)
         for tile_v in hl.tile(V, block_size=block_size_n):
             # Load log probabilities and convert to float32
             X = _input[tile_bt, tile_v]
